loging dialog component
